feat: combine input with ball physics
The input has been combined with ball physics. The player can shoot by dragging. The shot is made by converting the drag into speed. If the error with the perfect shooting speed is less than a certain threshold, then the perfect shot is made. The same logic is applied to shooting towards the billboard.